NetworkTest is working.  7 messages per second.  It should only get faster...
