+++ rust/src/main.rs
@@ -5,7 +5,7 @@ mod handlers;
 #[tokio::main]
 async fn main() -> flow::Result<()> {
     fdbserver::grv_master::foo();
-    flow::hello().await?;
+    handlers::hello().await?;
     println!("Goodbye, cruel world!");
 
     Ok(())
